Add `AnnotationsReader`.
