+++ index.ts
@@ -34,6 +34,9 @@ export * from "./src/Type/ReferenceType";
 export * from "./src/Type/DefinitionType";
 export * from "./src/Type/AnnotatedType";
 
+export * from "./src/AnnotationsReader";
+export * from "./src/AnnotationsReader/DefaultAnnotationsReader";
+
 export * from "./src/TypeFormatter";
 export * from "./src/SubTypeFormatter";
 export * from "./src/ChainTypeFormatter";
