+++ src/AnnotationsReader.ts
@@ -1,0 +1,6 @@
+import * as ts from "typescript";
+import { Annotations } from "./Type/AnnotatedType";
+
+export interface AnnotationsReader {
+    getAnnotations(node: ts.Node): Annotations;
+}
