+++ src/NodeParser/AnnotatedNodeParser.ts
@@ -3,36 +3,12 @@ import { Context } from "../NodeParser";
 import { SubNodeParser } from "../SubNodeParser";
 import { BaseType } from "../Type/BaseType";
 import { AnnotatedType, Annotations } from "../Type/AnnotatedType";
+import { AnnotationsReader } from "../AnnotationsReader";
 
 export class AnnotatedNodeParser implements SubNodeParser {
-    private static textTags: string[] = [
-        "title",
-        "description",
-
-        "format",
-        "pattern",
-    ];
-    private static jsonTags: string[] = [
-        "minimum",
-        "exclusiveMinimum",
-
-        "maximum",
-        "exclusiveMaximum",
-
-        "multipleOf",
-
-        "minLength",
-        "maxLength",
-
-        "minItems",
-        "maxItems",
-        "uniqueItems",
-
-        "default",
-    ];
-
     public constructor(
         private childNodeParser: SubNodeParser,
+        private annotationsReader: AnnotationsReader,
     ) {
     }
 
@@ -41,7 +17,7 @@ export class AnnotatedNodeParser implements SubNodeParser {
     }
     public createType(node: ts.Node, context: Context): BaseType {
         const baseType: BaseType = this.childNodeParser.createType(node, context);
-        const annotations: Annotations = this.parseAnnotations(this.getAnnotatedNode(node));
+        const annotations: Annotations = this.annotationsReader.getAnnotations(this.getAnnotatedNode(node));
         return !annotations ? baseType : new AnnotatedType(baseType, annotations);
     }
 
@@ -54,49 +30,6 @@ export class AnnotatedNodeParser implements SubNodeParser {
             return node.parent;
         } else {
             return node;
-        }
-    }
-
-    private parseAnnotations(node: ts.Node): Annotations {
-        const symbol: ts.Symbol = (node as any).symbol;
-        if (!symbol) {
-            return undefined;
-        }
-
-        const jsDocTags: ts.JSDocTagInfo[] = symbol.getJsDocTags();
-        if (!jsDocTags || !jsDocTags.length) {
-            return undefined;
-        }
-
-        const annotations: Annotations = jsDocTags.reduce((result: Annotations, jsDocTag: ts.JSDocTagInfo) => {
-            const value: any = this.parseJsDocTag(jsDocTag);
-            if (value !== undefined) {
-                result[jsDocTag.name] = value;
-            }
-
-            return result;
-        }, {});
-        return Object.keys(annotations).length ? annotations : undefined;
-    }
-
-    private parseJsDocTag(jsDocTag: ts.JSDocTagInfo): any {
-        if (!jsDocTag.text) {
-            return undefined;
-        }
-
-        if (AnnotatedNodeParser.textTags.indexOf(jsDocTag.name) >= 0) {
-            return jsDocTag.text;
-        } else if (AnnotatedNodeParser.jsonTags.indexOf(jsDocTag.name) >= 0) {
-            return this.parseJson(jsDocTag.text);
-        } else {
-            return undefined;
-        }
-    }
-    private parseJson(value: string): any {
-        try {
-            return JSON.parse(value);
-        } catch (e) {
-            return undefined;
         }
     }
 }
